deleted original profile page - moved players stuff to profile/edit page. Deleted gameplay link
