+++ client/src/pages/Profile.js
@@ -1,49 +1,143 @@
-import React from "react";
-import { Jumbotron } from "reactstrap";
+import React, { Component } from "react";
+import DeleteBtn from "../components/DeleteBtn";
+import Jumbotron from "../components/Jumbotron";
+import API from "../utils/API";
+import { Link } from "react-router-dom";
 import { Col, Row, Container } from "../components/Grid";
+import { List, ListItem } from "../components/List";
+import { Input, TextArea, FormBtn } from "../components/Form";
+import AvatarSetter from "../components/PlayerAvatar/AvatarSetter";
+import Navbar from "../components/Navbar/Navbar";
 import NewAvatar from "../components/PlayerAvatar/NewAvatar";
 
-const Profile = props => {
-  //console.log(props.user);
-  console.log(props.user.robot.RobotColor);
-  return (
-    <Container>
-      <br />
-      <Row>
-        <Col size="sm-1" />
-        <Col size="sm-10">
-          <Jumbotron className="text-center">
-            <h1 className="text-center">
-              <small>
-                <i className="fas fa-baseball-ball fa-spin" />
-              </small>{" "}
-              <strong> Player stats</strong>
-            </h1>
-            <p className="lead text-center"> Players</p>
-            <hr className="my-2 text-center" />
+class Players extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      players: [],
+      firstName: props.user.firstName,
+      robot: props.user.robot
+    };
+  }
 
-            {/* <Link to="/teams" className="btn btn-outline-primary">teams</Link> 
-        <Link to="/games" className="btn btn-outline-primary">games</Link>  */}
-            <div>
-              <NewAvatar
-                RobotAntenna={props.user.robot.RobotAntenna}
-                RobotColor={props.user.robot.RobotColor}
-                RobotImage={props.user.robot.RobotImage}
-              />
-              <br />
-              Name: {props.user.firstName} {props.user.lastName}
-              <br />
-              {/* Game Score: {props.user.gameScore}
-              <br />
-              Historical Score: {props.user.histScore}
-              <br /> */}
-            </div>
-          </Jumbotron>
-        </Col>
-        <Col size="sm-1" />
-      </Row>
-    </Container>
-  );
-};
+  componentDidMount() {
+    this.loadplayers();
+  }
 
-export default Profile;
+  loadplayers = () => {
+    API.getplayers()
+      .then(res => this.setState({ players: res.data }))
+      .catch(err => console.log(err));
+  };
+
+  deleteplayer = id => {
+    API.deleteplayer(id)
+      .then(res => this.loadplayers())
+      .catch(err => console.log(err));
+  };
+
+  handleInputChange = event => {
+    const { value } = event.target;
+    this.setState({
+      firstName: value
+    });
+  };
+
+  handleAvatarChange = robot => {
+    console.log(robot);
+    this.setState({
+      robot: robot
+    });
+  };
+
+  handleFormSubmit = event => {
+    event.preventDefault();
+    if (this.state.firstName) {
+      API.saveplayer(this.props.user._id, {
+        firstName: this.state.firstName,
+        robot: this.state.robot
+      })
+        .then(res => window.location.reload())
+        .catch(err => console.log(err));
+    }
+  };
+
+  render() {
+    return (
+      <Container fluid>
+        <Row>
+          <Col size="md-6">
+            <Jumbotron>
+              <h1>
+                <i class="fas fa-user-plus" /> <strong>Edit Player</strong>
+              </h1>
+
+              <div
+                style={{
+                  display: "inline-block",
+                  marginBottom: "10px"
+                }}
+              >
+                <AvatarSetter
+                  robot={this.props.user.robot}
+                  onChange={this.handleAvatarChange}
+                />
+              </div>
+              <form>
+                <Input
+                  value={this.state.firstName}
+                  onChange={this.handleInputChange}
+                  name="firstName"
+                  placeholder="First Name"
+                />
+
+                <FormBtn
+                  disabled={!this.state.firstName}
+                  onClick={this.handleFormSubmit}
+                >
+                  Save Changes
+                  {/* <button>
+                  {" "}
+                  Submit player <i class="fa fa-check-square-o" />
+                </button> */}
+                  {/* <i class="fas fa-check-square"></i> Submit Player */}
+                </FormBtn>
+              </form>
+            </Jumbotron>
+          </Col>
+          <Col size="md-6 sm-12">
+            <Jumbotron className="text-center">
+              <div
+                className="text-center"
+                style={{
+                  margin: "auto"
+                }}
+              >
+                <h1>
+                  <i class="fas fa-user-friends" /> <strong>Player Info</strong>
+                </h1>
+
+                <div
+                  style={{
+                    display: "inline-block"
+                  }}
+                >
+                  <NewAvatar
+                    RobotAntenna={this.props.user.robot.RobotAntenna}
+                    RobotColor={this.props.user.robot.RobotColor}
+                    RobotImage={this.props.user.robot.RobotImage}
+                  />
+                  <br />
+                  Name: {this.props.user.firstName} {this.props.user.lastName}
+                  <br />
+                </div>
+              </div>
+            </Jumbotron>
+          </Col>
+        </Row>
+      </Container>
+    );
+  }
+}
+
+export default Players;
